+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Current
 -------
 
 - Handle callable on API infos
+- Drop/merge flask_restful `flask_restful.RequestParser`
 
 0.8.5 (2015-12-12)
 ------------------